+++ scalding/src/main/scala/commbank/coppersmith/scalding/EavtTextSink.scala
@@ -1,3 +1,17 @@
+//
+// Copyright 2016 Commonwealth Bank of Australia
+//
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+//        http://www.apache.org/licenses/LICENSE-2.0
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+//
+
 package commbank.coppersmith.scalding
 
 import commbank.coppersmith.scalding.HiveSupport.{FailJob, DelimiterConflictStrategy}
@@ -8,10 +22,6 @@ import au.com.cba.omnia.maestro.api._, Maestro._
 
 import commbank.coppersmith.{FeatureValue, Feature}, Feature._, Value._
 import commbank.coppersmith.thrift.Eavt
-
-/**
-  * Created by vaughaew on 4/05/2016.
-  */
 
 object EavtTextSink {
   import TextSink._